// src/generated/model/programLevelRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { ProgramLevelRequestStatus } from './programLevelRequestStatus.ts';

export interface ProgramLevelRequest {
  programId: string;
  /**
     * @minLength 0
     * @maxLength 100
     */
  name: string;
  status: ProgramLevelRequestStatus;
  displayOrder?: number;
}