// src/generated/model/updateProgramLevelRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { UpdateProgramLevelRequestStatus } from './updateProgramLevelRequestStatus.ts';

export interface UpdateProgramLevelRequest {
  /**
     * @minLength 0
     * @maxLength 100
     */
  name: string;
  status: UpdateProgramLevelRequestStatus;
  displayOrder?: number;
}